requirements.txt
absl-py==0.9.0
addict==2.4.0
anykeystore==0.2
apex==0.9.10.dev0
#apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astroid==2.5
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.0
blinker==1.4
#Brlapi==0.7.0
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
chumpy==0.69
Click==7.0
colorama==0.4.3
#command-not-found==0.3
ConfigArgParse==1.5.3
configer==1.3.1
configparser==4.0.2
cryptacular==1.5.5
cryptography==2.8
cssselect==1.1.0
#cupshelpers==1.0
cycler==0.10.0
dataclasses==0.6
#dbus-python==1.2.16
decorator==4.4.1
#defer==1.0.6
defusedxml==0.6.0
distlib==0.3.3
distro==1.4.0
#distro-info===0.23ubuntu1
docopt==0.6.2
dotmap==1.3.25
#duplicity==0.8.12.0
easydict==1.9
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.0.12
freetype-py==2.2.0
future==0.18.2
fvcore==0.1.5.post20211015
gdown==3.13.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
h5py==2.10.0
h5py-cache==1.0
html5lib==1.0.1
httplib2==0.14.0
hupper==1.9.1
idna==2.8
imageio==2.6.1
importlib-metadata==1.4.0
iopath==0.1.9
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.7.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
#jupyter==1.0.0
#jupyter-client==5.3.4
#jupyter-console==6.0.0
#jupyter-core==4.6.1
#jupyterlab-pygments==0.1.2
#jupyterlab-widgets==1.0.0
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.1.0
kornia==0.5.0
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.5.2
lockfile==0.12.2
loguru==0.5.3
#louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==8.1.0
mpmath==1.2.1
nbclient==0.5.1
nbconvert==5.6.1
nbformat==5.0.3
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.4
notebook==6.0.2
numpy==1.21.2
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
open3d==0.12.0
opencv-python==4.5.2.52
OpenEXR==1.3.2
openpyxl==3.0.9
packaging==20.8
pandas==1.0.0
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
PasteDeploy==2.0.1
pbkdf2==1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==2.4.0
plotly==5.3.1
plyfile==0.7.3
portalocker==2.3.2
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
#pycups==1.9.73
pyglet==1.5.14
Pygments==2.5.2
#PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==2.4.6
pyramid==1.10.4
pyramid-mailer==0.15.1
pyrender==0.1.43
pyRFC3339==1.1
pyrsistent==0.15.7
PySocks==1.7.1
#python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.1
#python-debian===0.1.36ubuntu1
python3-openid==3.1.0
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
QtPy==1.9.0
reportlab==3.5.34
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
scour==0.37
#screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
smplx==0.1.28
soupsieve==1.9.5
SQLAlchemy==1.3.12
ssh-import-id==5.10
StereoVision==1.0.4
sympy==1.8
#system-service==0.3
systemd-python==234
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.0
torchgeometry==0.1.2
torchvision==0.9.0
tornado==6.0.3
tqdm==4.62.3
traitlets==4.3.3
transaction==3.0.0
translationstring==1.3
trimesh==3.9.33
typing-extensions==3.7.4.3
#ubuntu-advantage-tools==27.2
#ubuntu-drivers-common==0.0.0
#ufw==0.36
#unattended-upgrades==0.1
urllib3==1.25.7
#usb-creator==0.3.7
velruse==1.1.1
venusian==3.0.0
virtualenv==20.8.1
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.2.1
wtforms-recaptcha==0.3.2
#xkit==0.0.0
xmltodict==0.12.0
yacs==0.1.8
yapf==0.30.0
yarg==0.1.9
zipp==1.0.0
zope.deprecation==4.4.0
zope.interface==4.7.1
zope.sqlalchemy==1.2
